Improved error message when config not found
Badlist requires configuring sources in two places,
on the source list and in the service configuration.
If the name of the source isn't found in the service
config, the default KeyError is handled in the way
that both Update sources and logs shows the source name
as the error. This change handles the KeyError and re-raise
as ValueError with the message pointing user to the missing
configuration. It's especially important for users trying
to add their own sources, not included in the default config.